wip: create LogManager.cs and move logging logic in it
