+++ Source/VidcutterModule.cs
@@ -1,14 +1,8 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
-using System.Linq;
-using System.Net.Http;
-using System.Reflection;
-using System.Threading;
 using Celeste.Mod.UI;
 using Microsoft.Xna.Framework;
 using Monocle; 
@@ -35,9 +29,6 @@ public class VidcutterModule : EverestModule {
     public override Type SaveDataType => typeof(VidcutterModuleSaveData);
     public static VidcutterModuleSaveData SaveData => (VidcutterModuleSaveData)Instance._SaveData;
 
-    public static string logPath;
-    public static StreamWriter LogFileWriter = null;
-
     public static Vector2? previousRespawnPoint = null;
     public static Level previousLevel = null;
     public static bool processWhenClose = false;
@@ -52,68 +43,9 @@ public class VidcutterModule : EverestModule {
         Logger.SetLogLevel(nameof(VidcutterModule), LogLevel.Info);
     }
 
-    public static void Log(string message, Session session = null) {
-        string toLog = $"[{DateTime.Now:yyyy-MM-dd HH:mm:ss.fff}] ";
-        if (session != null) {
-            string sid = session.Area.SID;
-            if (sid.StartsWith("Celeste/")) {
-                sid = $"AREA_{sid.Substring(8, 1)}";
-                if (sid == "AREA_L") {
-                    sid = "AREA_10";
-                }
-            }
-            toLog += Dialog.Clean(sid);
-            if (session.Area.Mode.ToString().EndsWith("Side")) {
-                toLog += $" [{session.Area.Mode.ToString()[0]}-Side]";
-            }
-            toLog += $" | {session.Level} | ";
-        }
-        toLog += message;
-        LogFileWriter.WriteLine(toLog);
-    }
-
-    public static List<LoggedString> getAllLogs(string video, string level = null) {
-        DateTime startVideo = File.GetCreationTime(video);
-        TimeSpan? duration = VideoCreation.getVideoDuration(video);
-        Logger.Info("Vidcutter", $"Video {video} started at {startVideo} and has duration {duration}");
-        if (duration == null) {
-            return new List<LoggedString>();
-        }
-        DateTime endVideo = startVideo + (TimeSpan)duration;
-        return getAllLogs(startVideo, endVideo, level);
-    }
-
-    public static List<LoggedString> getAllLogs(DateTime? startVideo = null, DateTime? endVideo = null, string level = null) {
-        LogFileWriter.Close();
-        string[] lines = File.ReadAllLines(logPath);
-        List<LoggedString> parsedLines = new List<LoggedString>();
-        foreach (string line in lines) {
-            DateTime logTime = DateTime.Parse(line.Substring(1, 23));
-            string[] loggedEvent = line.Substring(26).Split(" | ");
-            bool condition = true;
-            if (startVideo != null) {
-                condition &= startVideo <= logTime;
-            }
-            if (endVideo != null) {
-                condition &= logTime <= endVideo;
-            }
-            if (level != null) {
-                condition &= loggedEvent[0] == level;
-            }
-            if (condition) {
-                parsedLines.Add(new LoggedString(logTime, loggedEvent[2], loggedEvent[0], loggedEvent[1]));
-            }
-        }
-
-        LogFileWriter = new StreamWriter(logPath, true) {
-            AutoFlush = true
-        };
-        return parsedLines;
-    }
-
     public static void OnComplete(On.Celeste.Level.orig_RegisterAreaComplete orig, Level self) {
         if (!self.Completed) {
-            Log("LEVEL COMPLETE", session: self.Session);
+            LogManager.Log("LEVEL COMPLETE", session: self.Session);
             processWhenClose = false;
         }
         orig(self);
@@ -121,7 +53,7 @@ public class VidcutterModule : EverestModule {
 
     public static void OnDeath(On.Celeste.Level.orig_LoadLevel orig, Level self, Player.IntroTypes playerIntro, bool isFromLoader = false) {
         if (playerIntro == Player.IntroTypes.Respawn) {
-            Log("DEATH", session: self.Session);
+            LogManager.Log("DEATH", session: self.Session);
             processWhenClose = false;
             inState = false;
         }
@@ -129,14 +61,14 @@ public class VidcutterModule : EverestModule {
     }
 
     public static void OnBegin(On.Celeste.Level.orig_Begin orig, Level self) {
-        Log("LEVEL LOADED", session: self.Session);
+        LogManager.Log("LEVEL LOADED", session: self.Session);
         inState = false;
         orig(self);
     }
 
     public static void OnCollectStrawberry(On.Celeste.Strawberry.orig_OnCollect orig, Strawberry self) {
         if (!inState) {
-            Log("ROOM PASSED", session: self.SceneAs<Level>().Session);
+            LogManager.Log("ROOM PASSED", session: self.SceneAs<Level>().Session);
         }
         orig(self);
     }
@@ -144,13 +76,13 @@ public class VidcutterModule : EverestModule {
     public static IEnumerator OnCollectCassette(On.Celeste.Cassette.orig_CollectRoutine orig, Cassette self, Player player) {
         yield return new SwapImmediately(orig(self, player));
         if (!inState) {
-            Log("ROOM PASSED", session: self.SceneAs<Level>().Session);
+            LogManager.Log("ROOM PASSED", session: self.SceneAs<Level>().Session);
         }
     }
 
     public static void OnRestart(On.Celeste.LevelExit.orig_ctor orig, LevelExit self, LevelExit.Mode mode, Session session, HiresSnow snow) {
         if (mode == LevelExit.Mode.Restart) {
-            Log("RESTART CHAPTER", session: session);
+            LogManager.Log("RESTART CHAPTER", session: session);
         }
         orig(self, mode, session, snow);
     }
@@ -165,14 +97,14 @@ public class VidcutterModule : EverestModule {
         if (previousRespawnPoint != respawnPoint || previousLevel != self.SceneAs<Level>()) {
             previousLevel = self.SceneAs<Level>();
             previousRespawnPoint = respawnPoint;
-            Log($"ROOM PASSED", session: self.SceneAs<Level>().Session);
+            LogManager.Log($"ROOM PASSED", session: self.SceneAs<Level>().Session);
             processWhenClose = true;
         }
         float deltaY = Math.Abs(playerPos.Y - respawnPoint.Value.Y);
         float deltaX = Math.Abs(playerPos.X - respawnPoint.Value.X);
         double distance = Math.Sqrt(deltaX * deltaX + deltaY * deltaY);
         if (distance <= 50 && processWhenClose && !inState) {
-            Log($"ROOM PASSED", session: self.SceneAs<Level>().Session);
+            LogManager.Log($"ROOM PASSED", session: self.SceneAs<Level>().Session);
             processWhenClose = false;
         }
     }
@@ -180,9 +112,9 @@ public class VidcutterModule : EverestModule {
     public static void onLoadState(Level level) {
         Vector2? playerPosition = level.Tracker.GetEntity<Player>()?.Position;
         if (playerPosition == level.Session.RespawnPoint) {
-            Log("DEATH", session: level.Session);
+            LogManager.Log("DEATH", session: level.Session);
         } else {   
-            Log("STATE", session: level.Session);
+            LogManager.Log("STATE", session: level.Session);
         }
         processWhenClose = false;
         previousRespawnPoint = level.Session.RespawnPoint;
@@ -236,8 +168,8 @@ public class VidcutterModule : EverestModule {
         if (!Directory.Exists(logFolder)) {
             Directory.CreateDirectory(logFolder);
         }
-        logPath = Path.Combine("./VidCutter/", Path.Combine("logs", "log.txt"));
-        LogFileWriter = new StreamWriter(logPath, true) {
+        LogManager.logPath = Path.Combine("./VidCutter/", Path.Combine("logs", "log.txt"));
+        LogManager.LogFileWriter = new StreamWriter(LogManager.logPath, true) {
             AutoFlush = true
         };
         On.Celeste.Level.RegisterAreaComplete += OnComplete;
@@ -286,7 +218,7 @@ public class VidcutterModule : EverestModule {
     }
 
     public override void Unload() {
-        LogFileWriter.Close();
+        LogManager.LogFileWriter.Close();
         On.Celeste.Level.RegisterAreaComplete -= OnComplete;
         On.Celeste.Level.Begin -= OnBegin;
         On.Celeste.Level.LoadLevel -= OnDeath;
@@ -297,35 +229,5 @@ public class VidcutterModule : EverestModule {
         if (SpeedrunToolInstalled) {
             VidcutterSpeedrunToolImport.Unregister(action);
         }
-    }
-
-    public static void deleteLogs(List<ProcessedVideo> rows){
-        List<LoggedString> allLogs = getAllLogs();
-        foreach (ProcessedVideo row in rows) {
-            string video = Path.Combine(Settings.VideoFolder, row.Video);
-            string level = row.Level;
-            DateTime startVideo = File.GetCreationTime(video);
-            TimeSpan? duration = VideoCreation.getVideoDuration(video);
-            if (duration == null) {
-                return;
-            }
-            DateTime endVideo = startVideo + (TimeSpan)duration;
-            List<LoggedString> allLogsCopy = [.. allLogs];
-            foreach (LoggedString log in allLogsCopy) {
-                if (startVideo < log.Time && log.Time < endVideo && log.Level == level) {
-                    allLogs.Remove(log);
-                }
-            }
-        }
-
-        LogFileWriter.Close();
-        using (StreamWriter writer = new StreamWriter(logPath, false)) {
-            foreach (LoggedString log in allLogs) {
-                writer.WriteLine(log.ToString());
-            }
-        }
-        LogFileWriter = new StreamWriter(logPath, true) {
-            AutoFlush = true
-        };
     }
 }
